refactor(api): remove unused authenticated health probe
