--- turbo/apps/api/src/signals/routes/health-auth-probe.ts
@@ -1,94 +1,0 @@
-import { initContract } from "@okouai/api-contracts/contracts/trpc-contract";
-import { command } from "ccstate";
-import { z } from "zod";
-
-import {
-  requiredAuthContext$,
-  setAuthContext$,
-  type AuthErrorResponse,
-} from "../auth/auth-context";
-import type { Capability } from "@okouai/api-contracts/contracts/capabilities";
-import type { AuthContext, AuthTokenType } from "../../types/auth";
-import type { RouteEntry } from "../route-entry";
-import { rawQuery$ } from "../context/hono";
-
-const c = initContract();
-
-export const healthAuthProbeContract = c.router({
-  check: {
-    method: "GET" as const,
-    path: "/health/auth",
-    headers: z.object({
-      authorization: z.string().optional(),
-      cookie: z.string().optional(),
-    }),
-    query: z.object({
-      acceptAnySandboxCapability: z.string().optional(),
-      requiredCapability: z.string().optional(),
-      accept: z.string().optional(),
-    }),
-    responses: {
-      200: z.unknown(),
-      401: z.object({
-        error: z.object({ message: z.string(), code: z.string() }),
-      }),
-      403: z.object({
-        error: z.object({ message: z.string(), code: z.string() }),
-      }),
-    },
-    summary: "Resolve auth context for the current request",
-  },
-});
-
-const probe$ = command(
-  async (
-    { get, set },
-    signal: AbortSignal,
-  ): Promise<
-    { readonly status: 200; readonly body: AuthContext } | AuthErrorResponse
-  > => {
-    const query = get(rawQuery$);
-
-    const options: {
-      requiredCapability?: Capability;
-      acceptAnySandboxCapability?: boolean;
-      accept?: readonly AuthTokenType[];
-    } = {};
-
-    if (query.acceptAnySandboxCapability === "true") {
-      options.acceptAnySandboxCapability = true;
-    }
-    if (query.requiredCapability) {
-      options.requiredCapability = query.requiredCapability as Capability;
-    }
-    if (query.accept) {
-      options.accept = query.accept.split(",") as AuthTokenType[];
-    }
-
-    const result = await set(requiredAuthContext$, options, signal);
-    if ("status" in result) {
-      return result;
-    }
-
-    // Post-filter: reject token types not in the accept list
-    if (options.accept && !options.accept.includes(result.tokenType)) {
-      return {
-        status: 403 as const,
-        body: {
-          error: {
-            message:
-              "This endpoint does not accept the provided credential type",
-            code: "FORBIDDEN",
-          },
-        },
-      };
-    }
-
-    set(setAuthContext$, result);
-    return { status: 200 as const, body: result };
-  },
-);
-
-export const healthAuthProbeRoutes: readonly RouteEntry[] = [
-  { route: healthAuthProbeContract.check, handler: probe$ },
-];
