+++ turbo/packages/api-contracts/src/contracts/__tests__/health.test.ts
@@ -1,21 +1,11 @@
 import { describe, expect, it } from "vitest";
 
-import {
-  healthAuthContract,
-  healthContract,
-  healthResponseSchema,
-} from "../health";
+import { healthContract, healthResponseSchema } from "../health";
 
 describe("health contract", () => {
   it("defines the API health endpoint", () => {
     expect(healthContract.check.method).toBe("GET");
     expect(healthContract.check.path).toBe("/health");
-  });
-
-  it("defines the authenticated API health endpoint", () => {
-    expect(healthAuthContract.check.method).toBe("GET");
-    expect(healthAuthContract.check.path).toBe("/health/auth");
-    expect(healthAuthContract.check.responses[401]).toBeDefined();
   });
 
   it("accepts the health response body", () => {
